Add Rancher image customization option
Added the ability for users to specify a custom Rancher image in the configuration. This change includes associating the custom image with the "rancherImage" variable in the host and tenant modules, and adding corresponding updates in the test and tools scripts to handle this option. Instructions are also updated in the README file.