Update software versions in several Terraform and Go modules
This update changes the version of required providers in Terraform modules for helm and aws, as well as the Go version in several Go modules. It also includes additional dependencies in the go.sum file. The goal is to take advantage of new features or fixes in these later versions to enhance performance and stability of the project.